crypto: Deprecate rand_bytes/1
OpenSSL has deprecated the function RAND_pseudo_bytes used by
crypto:rand_bytes/1, so this function is now deprecated in OTP too.

rand_bytes/3 also used this function, but was not documented
so we can remove it right away.

This commit also removes the fallback in generate_key to use
rand_bytes/1 if strong_rand_bytes/1 throws low entropy.
This is a potential incompatility but we think it is desirable
as crypto should provide cryptographically secure functions.