ssl: ERL-738 - Correct alert handling with new TLS sender process
With the new TLS sender process, solving ERL-622, alerts sent in the
connection state must be encrypted and sent by the TLS sender
process. This was not always honored for CLOSE alerts which could
result in an unexpected message as the CLOSE alert was encrypted using
an incorrect encryption state.