ssl: ERL-738 - Correct alert handling with new TLS sender process
With the new TLS sender process, solving ERL-622, TLS ALERTs sent in
the connection state must be encrypted and sent by the TLS sender
process. This to make sure that the correct encryption state is used
to encode the ALERTS. Care must also be taken to ensure a gracefull
close down behavior both for normal shutdown and downgrading from TLS
to tcp.

The original TR ERL-738 is verifyed by cowboy tests, and close down
behaviour by our tests. However we alas have not been able to yet
create a minimal test case for the originating problem.

Also it seems it has become less likly that we run in to the
tcp delivery problem, that is the guarantee is only on transport level,
not application level. Keep work around function in ssl_test_lib but
we can have better test as long as we do not get to much wobbling tests.